# internal/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API request bodies and responses are usually JSON-encoded, except for binary data and some of the OAuth2 endpoints. Note that the `Content-Type` header must be set to `application/json` for this, as the request may otherwise fail with error 400.\nWhen clients have a valid access token, e.g. obtained through the `POST /api/v1/session` or `POST /api/v1/oauth/token` endpoint, they can use a standard Bearer Authorization header to authenticate their requests. Submitting the access token with a custom `X-Auth-Token` header is supported as well.",
        "title": "PhotoPrism API",
        "contact": {},
        "version": "v1"
    },
    "host": "demo.photoprism.app",
    "paths": {
...
        "/api/v1/albums/{uid}/links": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links",
                    "Albums"
                ],
                "summary": "returns all share links for the given UID as JSON",
                "operationId": "GetAlbumLinks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "album uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Link"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links",
                    "Albums"
                ],
                "summary": "adds a new album share link and return it as JSON",
                "operationId": "CreateAlbumLink",
                "parameters": [
                    {
                        "type": "string",
                        "description": "album uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "link properties (currently supported: slug, expires)",
                        "name": "link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.Link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Link"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/albums/{uid}/links/{linkuid}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links",
                    "Albums"
                ],
                "summary": "updates an album share link and return it as JSON",
                "operationId": "UpdateAlbumLink",
                "parameters": [
                    {
                        "type": "string",
                        "description": "album uid",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "link uid",
                        "name": "linkuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "properties to be updated (currently supported: slug, expires, token)",
                        "name": "link",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.Link"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Link"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Links",
                    "Albums"
                ],
                "summary": "deletes an album share link",
                "operationId": "DeleteAlbumLink",
                "parameters": [
                    {
                        "type": "string",
                        "description": "album",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "link uid",
                        "name": "linkuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Link"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/i18n.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Link": {
            "type": "object",
            "properties": {
                "Comment": {
                    "type": "string"
                },
                "CreatedAt": {
                    "type": "string"
                },
                "CreatedBy": {
                    "type": "string"
                },
                "Expires": {
                    "type": "integer"
                },
                "MaxViews": {
                    "type": "integer"
                },
                "ModifiedAt": {
                    "type": "string"
                },
                "Perm": {
                    "type": "integer"
                },
                "ShareUID": {
                    "type": "string"
                },
                "Slug": {
                    "type": "string"
                },
                "Token": {
                    "type": "string"
                },
                "UID": {
                    "type": "string"
                },
                "VerifyPassword": {
                    "type": "boolean"
                },
                "Views": {
                    "type": "integer"
                }
            }
        },
...
        "form.Link": {
            "type": "object",
            "properties": {
                "CanComment": {
                    "type": "boolean"
                },
                "CanEdit": {
                    "type": "boolean"
                },
                "Expires": {
                    "type": "integer"
                },
                "MaxViews": {
                    "type": "integer"
                },
                "Password": {
                    "type": "string"
                },
                "Slug": {
                    "type": "string"
                },
                "Token": {
                    "type": "string"
                }
            }
        },
...
        "i18n.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}